most important stuff aligned from brif
compose files, nginx services, env sharing, tests refactoring
