+++ .env.example
@@ -11,7 +11,7 @@ DB_PASSWORD=somevalue
 PGADMIN_DEFAULT_EMAIL=somevalue
 PGADMIN_DEFAULT_PASSWORD=somevalue
 
-APP_PORT=somevalue
-APP_BASE_URL=example.com
-# local settings
-# APP_BASE_URL=http://localhost
+APP_PORT=8000
+APP_SUBDOMAIN="my-project"
+BASE_URL=example.com
+SECRET_KEY="superLongSecret*orNot"
