+++ app/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "papel"
-version = "0.1.0"
+version = "0.2.0"
 description = "Docker based tool to deploy APIs and data pipelines."
 authors = ["Pierre-Edouard Barrault"]
 
@@ -29,5 +29,7 @@ filterwarnings = [
     "error",
     "ignore::UserWarning",
     # temp fix for starlette/python3.8/Jinja3.1
-    "ignore::DeprecationWarning",
+    # "ignore::DeprecationWarning",
+    # TODO: remove once Jinja is updated
+    "ignore:'contextfunction' is renamed to 'pass_context', the old name will be removed in Jinja 3.1.:DeprecationWarning",
 ]
