--- pyproject.toml
@@ -1,89 +1,0 @@
-[tool.poetry]
-name = "papel-dev"
-version = "0.1.0"
-description = "Papel dev environment."
-authors = ["Pierre-Edouard Barrault <pierrz@tuta.io>"]
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.9"   # pinned after fastapi docker image
-aiofiles = "==0.7.0"
-amqp = "==5.0.6"
-anyio = "==3.3.4"
-astroid = "==2.8.5"
-attrs = "==21.2.0"
-"backports.entry-points-selectable" = "==1.1.1"
-billiard = "==3.6.4.0"
-celery = "==5.2.0"
-certifi = "==2021.10.8"
-cfgv = "==3.3.1"
-charset-normalizer = "==2.0.7"
-click = "==8.0.3"
-click-didyoumean = "==0.3.0"
-click-plugins = "==1.1.1"
-click-repl = "==0.2.0"
-distlib = "==0.3.3"
-fastapi = "==0.70.0"
-filelock = "==3.4.0"
-identify = "==2.3.6"
-idna = "==3.3"
-iniconfig = "==1.1.1"
-# TODO: check jinja update after 3.0
-Jinja2 = "==3.0.3"
-kombu = "==5.2.1"
-lazy-object-proxy = "==1.6.0"
-MarkupSafe = "==2.0.1"
-mccabe = "==0.6.1"
-mypy-extensions = "==0.4.3"
-nodeenv = "==1.6.0"
-numpy = "==1.21.4"
-packaging = "==21.2"
-pathspec = "==0.9.0"
-pip = "==21.2.4"
-platformdirs = "==2.4.0"
-pluggy = "==1.0.0"
-prompt-toolkit = "==3.0.22"
-psycopg2-binary = "==2.9.2"
-py = "==1.11.0"
-pycodestyle = "==2.8.0"
-pydantic = "==1.8.2"
-pyflakes = "==2.4.0"
-pyparsing = "==2.4.7"
-python-dotenv = "==0.19.2"
-pytz = "==2021.3"
-PyYAML = "==6.0"
-regex = "==2021.11.10"
-requests = "==2.26.0"
-setuptools = "==58.1.0"
-six = "==1.16.0"
-sniffio = "==1.2.0"
-starlette = "==0.16.0"
-toml = "==0.10.2"
-tomli = "==1.2.2"
-typing = "==3.7.4.3"
-typing_extensions = "==4.0.0"
-urllib3 = "==1.26.7"
-vine = "==5.0.0"
-virtualenv = "==20.10.0"
-wcwidth = "==0.2.5"
-wheel = "==0.37.0"
-wrapt = "==1.13.3"
-
-[tool.poetry.dev-dependencies]
-black = "==21.10b0"
-flake8 = "==4.0.1"
-isort = "==5.10.1"
-pre-commit = "==2.15.0"
-pylint = "==2.11.1"
-pytest = "==6.2.5"
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
-
-[tool.pytest.ini_options]
-filterwarnings = [
-    "error",
-    "ignore::UserWarning",
-    # temp fix for starlette/python3.8/Jinja3.1
-    "ignore::DeprecationWarning",
-]
