+++ setup/nginx/docker_base.conf
